+++ kwcoco/__init__.py
@@ -269,7 +269,7 @@ Testing:
 
 """
 
-__version__ = '0.5.4'
+__version__ = '0.5.5'
 
 
 __submodules__ = {