+++ kwcoco/cli/coco_split.py
@@ -19,11 +19,18 @@ class CocoSplitCLI(object):
         """
         default = {
             'src': scfg.Value(None, help='input dataset to split', position=1),
-            'dst1': scfg.Value('split1.mscoco.json', help='output path1'),
-            'dst2': scfg.Value('split2.mscoco.json', help='output path2'),
-            'factor': scfg.Value(3, help='ratio of items put in dset1 vs dset2'),
-            'rng': scfg.Value(None, help='random seed'),
+            'dst1': scfg.Value('split1.kwcoco.json', help='output path of the larger split'),
+            'dst2': scfg.Value('split2.kwcoco.json', help='output path of the smaller split'),
+            'factor': scfg.Value(3, help='number of items in dset1 for each item in dset2. Also defines the maximum number of splits that could be written.'),
+            'rng': scfg.Value(None, help='A random seed for reproducible splits'),
             'balance_categories': scfg.Value(True, help='if True tries to balance annotation categories across splits'),
+            'num_write': scfg.Value(1, isflag=True, help=ub.paragraph(
+                '''
+                The number of splits to write. Can be between 1 and ``factor``.
+                In the case that ``num_write > ``, then dst1 and dst2 datasets
+                must contain a {} format string specifier so each of the output
+                filesnames can be indexed.
+                ''')),
             'splitter': scfg.Value(
                 'auto', help=ub.paragraph(
                     '''
@@ -31,11 +38,14 @@ class CocoSplitCLI(object):
                     Using "image" will randomly assign each image to a partition.
                     Using "video" will randomly assign each video to a partition.
                     Using "auto" chooses "video" if there are any, otherwise "image".
-                    '''), choices=['auto', 'image', 'video'])
+                    '''), choices=['auto', 'image', 'video']),
+            'compress': scfg.Value('auto', help='if True writes results with compression'),
         }
         epilog = """
         Example Usage:
-            kwcoco split --src special:shapes8 --dst1=learn.mscoco.json --dst2=test.mscoco.json --factor=3 --rng=42
+            kwcoco split --src special:shapes8 --dst1=learn.kwcoco.json --dst2=test.kwcoco.json --factor=3 --rng=42
+
+            kwcoco split --src special:shapes8 --dst1="train_{03:d}.kwcoco.json" --dst2="vali_{0:3d}.kwcoco.json" --factor=3 --rng=42
         """
 
     @classmethod
@@ -61,6 +71,15 @@ class CocoSplitCLI(object):
 
         if config['src'] is None:
             raise Exception('must specify source: {}'.format(config['src']))
+
+        if config['num_write'] > 1:
+            if not set(str(config['dst1'])).issuperset(set('{}')):
+                raise Exception(
+                    'when num_write is True dst1 and dst2 must contain a {} format string placeholder')
+
+            if not set(str(config['dst2'])).issuperset(set('{}')):
+                raise Exception(
+                    'when num_write is True dst1 and dst2 must contain a {} format string placeholder')
 
         print('reading fpath = {!r}'.format(config['src']))
         dset = kwcoco.CocoDataset.coerce(config['src'])
@@ -101,7 +120,8 @@ class CocoSplitCLI(object):
         rng = kwarray.ensure_rng(config['rng'])
 
         shuffle = rng is not None
-        self = util_sklearn.StratifiedGroupKFold(n_splits=config['factor'],
+        factor = config['factor']
+        self = util_sklearn.StratifiedGroupKFold(n_splits=factor,
                                                  random_state=rng,
                                                  shuffle=shuffle)
 
@@ -109,21 +129,32 @@ class CocoSplitCLI(object):
             split_idxs = list(self.split(X=final_group_gids, y=final_group_cids, groups=final_group_ids))
         else:
             split_idxs = list(self.split(X=final_group_gids, y=final_group_gids, groups=final_group_ids))
-        idxs1, idxs2 = split_idxs[0]
 
-        gids1 = sorted(ub.unique(ub.take(final_group_gids, idxs1)))
-        gids2 = sorted(ub.unique(ub.take(final_group_gids, idxs2)))
+        dumpkw = {
+            'newlines': True,
+            'compress': config['compress'],
+        }
+        for split_num, (idxs1, idxs2) in enumerate(split_idxs):
+            print(f'Build split {split_num} / {factor} with ratio {len(idxs1)}:{len(idxs2)}')
+            idxs1, idxs2 = split_idxs[0]
+            gids1 = sorted(ub.unique(ub.take(final_group_gids, idxs1)))
+            gids2 = sorted(ub.unique(ub.take(final_group_gids, idxs2)))
 
-        dset1 = dset.subset(gids1)
-        dset2 = dset.subset(gids2)
+            dset1 = dset.subset(gids1)
+            dset2 = dset.subset(gids2)
+            print('stats(dset1): ' + ub.urepr(dset1.basic_stats(), nl=0))
+            print('stats(dset2): ' + ub.urepr(dset2.basic_stats(), nl=0))
 
-        dset1.fpath = config['dst1']
-        print('Writing dset1 = {!r}'.format(dset1.fpath))
-        dset1.dump(dset1.fpath, newlines=True)
+            dset1.fpath = str(config['dst1']).format(split_num)
+            dset2.fpath = str(config['dst2']).format(split_num)
+            print(f'Writing dset1({split_num} / {factor}) = {dset1.fpath!r}')
+            dset1.dump(**dumpkw)
+            print(f'Writing dset2({split_num} / {factor}) = {dset2.fpath!r}')
+            dset2.dump(**dumpkw)
 
-        dset2.fpath = config['dst2']
-        print('Writing dset2 = {!r}'.format(dset2.fpath))
-        dset2.dump(dset2.fpath, newlines=True)
+            if split_num + 1 >= config['num_write']:
+                break
+
 
 _CLI = CocoSplitCLI