+++ tests/test_compressed.py
@@ -15,3 +15,16 @@ def test_read_zipped_kwcoco():
     dset2 = kwcoco.CocoDataset(zip_fpath)
     assert dset2.dataset == dset.dataset
     assert dset2.dataset is not dset.dataset
+
+
+def test_compress_dump():
+    import tempfile
+    import kwcoco
+    self = kwcoco.CocoDataset.demo()
+    file = tempfile.NamedTemporaryFile('wb')
+    self.dump(file, compress=True)
+    import zipfile
+    assert zipfile.is_zipfile(file.name)
+    self2 = kwcoco.CocoDataset(file.name)
+    assert self2.dataset == self.dataset
+    assert self2.dataset is not self.dataset